requirements.txt
web3==6.11.0
python-dotenv==1.0.0
rich==13.6.0
setuptools==68.2.2